Hackathon/Image_chatbot_website/requirements.txt
flask
tensorflow
numpy
pillow
requests
gunicorn
python-dotenv
